before:

`pixie .............................. 0.934 ms      0.945 ms    ±0.011  x1000`

after:

`pixie .............................. 0.498 ms      0.502 ms    ±0.002  x1000`